requirements.txt
qasync
asyncify
requests
MaaFw >= 4.1.0
psutil
PySide6-Fluent-Widgets
wmi
cryptography
